+++ Blocks5/src/cf_blend.cpp
@@ -9,10 +9,10 @@ CF_Blend::~CF_Blend()
 {
 }
 
-void CF_Blend::render(double t,
+void CF_Blend::render(float t,
 					  uint oldImageID,
 					  uint newImageID)
 {
 	// the old image, fading out over the new one already on the screen
-	drawImage(oldImageID, Vec4f(1.0f, 1.0f, 1.0f, static_cast<float>(1.0 - t)));
+	drawImage(oldImageID, Vec4f(1.0f, 1.0f, 1.0f, 1.0f - t));
 }
